+++ addons/point_of_sale/views/pos_templates.xml
@@ -2,6 +2,9 @@
 <odoo>
 <template id="pos_assets_backend" name="POS UI Backend Assets" inherit_id="web.assets_backend" primary="True">
     <xpath expr="//script[contains(@src, '/mail/static/src/js/services/mail_service.js')]" position="replace"></xpath>
+    <xpath expr="." position="inside">
+        <script type="text/javascript" src="/point_of_sale/static/src/js/owl_pos_root.js"></script>
+    </xpath>
 </template>
 
 <template id="index" name="POS Index">&lt;!DOCTYPE html&gt;
